Add tests for merge-adjacent-kpi-blocks Rename number.ts to wrap-around-length.ts (and also the associated spec)
